functions/orchestrator/requirements.txt
functions-framework==3.9.2
requests==2.32.3
google-auth==2.23.0